+++ lib/exwiw/table_config.rb
@@ -10,6 +10,7 @@ module Exwiw
     attribute :belongs_tos, array(BelongsTo)
     attribute :columns, array(TableColumn)
     attribute :bulk_insert_chunk_size, optional(Integer), skip_serializing_if_nil: true
+    attribute :skip, Serdes::OptionalType.new(Serdes::ConcreteType.new(Boolean)), skip_serializing_if_nil: true
 
     def self.from_symbol_keys(hash)
       from(JSON.parse(hash.to_json))
@@ -76,6 +77,7 @@ module Exwiw
         merged_table.filter = filter
         merged_table.belongs_tos = passed_table.belongs_tos
         merged_table.bulk_insert_chunk_size = passed_table.bulk_insert_chunk_size
+        merged_table.skip = skip
 
         receiver_column_by_name = columns.each_with_object({}) { |column, hash| hash[column.name] = column }